Move wildcardToPattern from PaneBuilder to Utils
